Fixed everything I did yesterday
Added 80% external tool functionality, created hitbox to follow character around and register collision (when the time comes)

+++ ExternalTool_CharacterEditor/ExternalTool_CharacterEditor/Form1.cs
@@ -15,7 +15,7 @@ namespace ExternalTool_CharacterEditor
     {
 
         // attributes: easier to save to files.
-        public float pSpeed, eSpeed;
+        public int pSpeed, eSpeed;
         public int pDamage, eDamage;
         public int pHealth, eHealth;
         public int pJumpHeight, eJumpHeight;
